🎨 adds plan action and bootstraps to default store to handle plans
