+++ Controller/Adminhtml/Plans/Create.php
@@ -2,39 +2,12 @@
 
 namespace MundiPagg\MundiPagg\Controller\Adminhtml\Plans;
 
-use Magento\Backend\App\Action;
-use Magento\Backend\App\Action\Context;
-use Magento\Framework\Registry;
-use Magento\Framework\View\Result\PageFactory;
+use MundiPagg\MundiPagg\Controller\Adminhtml\Plans\PlanAction;
 use Mundipagg\Core\Recurrence\Aggregates\Plan;
-use Mundipagg\Core\Recurrence\Repositories\PlanRepository;
 use Mundipagg\Core\Recurrence\Services\PlanService;
 
-class Create extends Action
+class Create extends PlanAction
 {
-    protected $resultPageFactory = false;
-    /**
-     * @var ProductsPlanFactory
-     */
-    private $productsPlanFactory;
-
-    /**
-     * Constructor
-     *
-     * @param Context $context
-     * @param PageFactory $resultPageFactory
-     * @param Registry $coreRegistry
-     */
-    public function __construct(
-        Context $context,
-        PageFactory $resultPageFactory,
-        Registry $coreRegistry
-    ) {
-        parent::__construct($context);
-        $this->resultPageFactory = $resultPageFactory;
-        $this->coreRegistry = $coreRegistry;
-    }
-
     /**
      * Index action
      *
@@ -43,7 +16,7 @@ class Create extends Action
     public function execute()
     {
         $planId = (int) $this->getRequest()->getParam('id');
-        if($planId) {
+        if ($planId) {
             //@todo this should be a product plan core object
             $planService = new PlanService();
             $planData = $planService->findById($planId);
